cleanups, add options
